Convert from a purely CPython C extension to a mix of C and ctypes
This represents a rewrite of most of the library. It reduces the amount of C
required by about two thirds, reducing the potential for crashing and memory
leak bugs. It adds Python references to Lua objects, making it more beneficial
to start one Lua VM and set of compiled code and reuse it. And it makes the
garbage collector interaction much smarter.