Implement library sandboxing differently
This is a breaking API change that should really have been done in 2.0.0. It
renames SimpleSandboxedExecutor, changes the call structure for sandboxing,
and changes the expected format of the sandboxer lua code. This also breaks
luajit compatibility.

A result is that the tests have a shim to look more like the old call
structure. This should probably be revisited so the tests aren't so divorced
from reality

Also fixes some simple bugs with TypeErrors