Try to be smarter about when we release the GIL.
ctypes by default releases the GIL when calling out to any C function referred
by `ctypes.CDLL`. But we know that some of the functions that we call are very
fast and primarily CPU or allocation bound (like `lua_pushnumber`) and some of
them can take arbitrary amounts of time (like `lua_pcallk`). Releasing the GIL
only to instantly reacquire it isn't free, so we can use what we know about
individual functions to be more choosy about when we release it. We do this
using `ctypes.PyDLL`, which is like CDLL except it doesn't release the GIL and
checks for a Python exception after the function returns.

perf before:

    capsule_test 19.2012419701
    limiter_test 25.3112540245
    re_test 30.5804388523
    simple_test 22.9568750858

perf after:

    capsule_test 18.3725008965
    limiter_test 24.7543950081
    re_test 29.3595330715
    simple_test 22.0168280602

The difference isn't huge but is a pretty consistent 4%-ish increase across the
board in these simple, short single-threaded tests. That number may fall as the
complexity of the lua code is increased or rise as the number of Capsule/Python
function accesses increase.

I suspect but don't have numbers to indicate that the difference is larger in a
multi-threaded environment.

Closes https://github.com/ketralnis/lua_sandbox/issues/3