Capsules cache their lookups
Capsules are objects to opaquely wrap Python objects and pass them to
Lua. If you make a Capsule(dict), Lua code can get the keys of that dict
lazily. That lets us pass large objects to Lua and only pay for copying
in the data that's actually requested.

The problem is, if you do that then it's copied every time it's asked
for. This patch adds a cache so that if that key has already been
copied into Lua, we don't have to copy it again the second time (and can
even avoid having to grab the GIL to do it)