Add an interface to raw Lua strings
Normally when passing data back and forth between Pytho and Lua we just
copy it every time. But sometimes that can be really expensive,
particularly if it's the same data over and over. This adds an interface
to get zero-copy access to the Lua strings using the lua_tolstring
interface.

Python functions can't always deal with these buffer objects directly
and so this isn't necessarily always free. But it's useful for re.search
and co, who can deal with them just fine.