Performance improvements by using the registry less
Using the Lua C library, it's not possible to hold a Lua value in a
variable (and by extension for us a Python variable). Instead all Lua
values are stored on the Lua stack, and you can keep track of them via
the integer on the stack. That works okay for small self-contained
functions that take and return values, but then how do you keep
long-lived references to compiled functions, or return values back to
Python where they may outlive the function that created them or be
stored in containers?

The answer is that Lua also has another global place to store things
called the Registry. It has a call `luaL_ref` which allocates and
returns an index into the registry and promises that the registry counts
as a reference to the value so it won't be GC'd

So every time lua_sandbox needs to talk about a Lua value it wraps it in
a Python object called a `LuaValue` which maintains the registry index
and manages moving the Lua value in and out of it. That works, but it
means that we have a surprising number of calls into Lua to do really
basic operations. In particular `LuaValue.__setitem__`, called when
setting sandbox globals like `lua.sandbox['thing'] = Something()`,
accounts for a lot of time in my own use case for lua_sandbox.
`LuaValue.__setitem__` does a lot of stuff including serialising
from Python to Lua and actually updating the Lua table, but some
fraction of that time is spent just moving stuff in and out of the
registry. Not a lot, but enough to show up in profiling.

This patch tries to cut that down a bit by cutting out some of the
registry management down. It does that by splitting LuaValue into
StackValue (which now represents a Lua value on the stack) and
RegistryValue (which does what LuaValue was doing before with all of the
registry management). To keep the external API stable we only use
StackValue internally and continue to use RegistryValue everywhere that
we leak references into our callers so it's still safe for them to keep
the references that we give them